+++ lib/src/marker_cluster_layer_options.dart
@@ -3,6 +3,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:flutter_map/plugin_api.dart';
+import 'package:flutter_map_marker_cluster/src/node/marker_cluster_node.dart';
 import 'package:flutter_map_marker_popup/extension_api.dart';
 
 class PolygonOptions {
@@ -108,6 +109,9 @@ class MarkerClusterLayerOptions extends LayerOptions {
   /// Function to call when markers are clustered
   final void Function(List<Marker>) onMarkersClustered;
 
+  /// Function to call when a cluster Marker is tapped
+  final void Function(MarkerClusterNode) onClusterTap;
+
   /// Popup's options that show when tapping markers or via the PopupController.
   final PopupOptions popupOptions;
 
@@ -130,6 +134,7 @@ class MarkerClusterLayerOptions extends LayerOptions {
     this.polygonOptions = const PolygonOptions(),
     this.showPolygon = true,
     this.onMarkerTap,
+    this.onClusterTap,
     this.onMarkersClustered,
     this.popupOptions,
   }) : assert(builder != null);